+++ R/RcppExports.R
@@ -5,24 +5,24 @@ impute_conj_mvnorm_mu_cpp <- function(y, mu, impute, Q, mu0, tau0) {
     .Call('_gibbs_utils_impute_conj_mvnorm_mu_cpp', PACKAGE = 'gibbs.utils', y, mu, impute, Q, mu0, tau0)
 }
 
-m_binom <- function(p, proposal, k, n, mean, precision) {
-    .Call('_gibbs_utils_m_binom', PACKAGE = 'gibbs.utils', p, proposal, k, n, mean, precision)
+m_binom <- function(p, proposal, k, n, mean, precision, use_norm, norm) {
+    .Call('_gibbs_utils_m_binom', PACKAGE = 'gibbs.utils', p, proposal, k, n, mean, precision, use_norm, norm)
 }
 
-m_binom_mv <- function(p, proposal, k, n, mean, Q) {
-    .Call('_gibbs_utils_m_binom_mv', PACKAGE = 'gibbs.utils', p, proposal, k, n, mean, Q)
+m_binom_mv <- function(p, proposal, k, n, mean, Q, use_norm, norm) {
+    .Call('_gibbs_utils_m_binom_mv', PACKAGE = 'gibbs.utils', p, proposal, k, n, mean, Q, use_norm, norm)
 }
 
 qt_binom_approx <- function(around, k, n, mean, tau) {
     .Call('_gibbs_utils_qt_binom_approx', PACKAGE = 'gibbs.utils', around, k, n, mean, tau)
 }
 
-qt_binom <- function(p, k, n, mean, precision) {
-    .Call('_gibbs_utils_qt_binom', PACKAGE = 'gibbs.utils', p, k, n, mean, precision)
+qt_binom <- function(p, k, n, mean, precision, use_norm, norm) {
+    .Call('_gibbs_utils_qt_binom', PACKAGE = 'gibbs.utils', p, k, n, mean, precision, use_norm, norm)
 }
 
-qt_binom_mv <- function(p, k, n, mean, Q) {
-    .Call('_gibbs_utils_qt_binom_mv', PACKAGE = 'gibbs.utils', p, k, n, mean, Q)
+qt_binom_mv <- function(p, k, n, mean, Q, use_norm, norm) {
+    .Call('_gibbs_utils_qt_binom_mv', PACKAGE = 'gibbs.utils', p, k, n, mean, Q, use_norm, norm)
 }
 
 one_binom_slice <- function(p, k, n, mean, precision, w, nexpand, ncontract) {