+++ R/slice_sample.R
@@ -52,8 +52,10 @@ ss_pois_reg <- function(L, k, mean, precision, ..., w = 1, nexpand = 10, ncontra
 #'   \code{ss_binom_reg} slice samples and \code{mh_binom_reg} Metropolis-samples
 #'   \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
 #'   where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.
-#'   In the case that \code{n} is zero (or a row of \code{n} is zero in the multivariate case),
-#'   slice sampling is ignored in favor of the conjugate normal draw.
+#'
+#'   In the case that \code{n} is zero, slice sampling is ignored in favor of a normal draw.
+#'   In the special case when an entire (multivariate) row of \code{n} is zero, the entire row
+#'   is simultaneously drawn; when only some elements are zero, they're drawn univariately.
 #'
 #'   Both vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
 #'   \code{p} is assumed to be multivariately distributed, and different internals are used.