+++ src/mh_bin.cpp
@@ -13,10 +13,19 @@ bool accept_reject(double ratio) {
 }
 
 // [[Rcpp::export]]
-NumericVector m_binom(NumericVector p, NumericVector proposal, NumericVector k, NumericVector n, NumericVector mean, NumericVector precision) {
+NumericVector m_binom(NumericVector p, NumericVector proposal, NumericVector k, NumericVector n, NumericVector mean, NumericVector precision,
+                      LogicalVector use_norm, NumericVector norm) {
   NumericVector out = clone(p);
   LogicalVector accept(p.size());
+
+  int nrm = 0;
   for(int i=0; i < p.size(); i++) {
+    if(use_norm[i]) {
+      out[i]= norm[nrm++];
+      accept[i] = true;
+      continue;
+    }
+
     double ratio = one_m_ratio(p[i], proposal[i], k[i], n[i], mean[i], precision[i]);
     bool a = accept_reject(ratio);
     accept[i] = a;
@@ -31,16 +40,33 @@ NumericVector m_binom(NumericVector p, NumericVector proposal, NumericVector k, 
 
 
 // [[Rcpp::export]]
-NumericVector m_binom_mv(NumericMatrix p, NumericMatrix proposal, NumericMatrix k, NumericMatrix n, NumericMatrix mean, NumericMatrix Q) {
+NumericVector m_binom_mv(NumericMatrix p, NumericMatrix proposal, NumericMatrix k, NumericMatrix n, NumericMatrix mean, NumericMatrix Q,
+                         LogicalVector use_norm, NumericMatrix norm) {
   NumericMatrix out = clone(p);
   LogicalMatrix accept(p.nrow(), p.ncol());
+
+  int nrm = 0;
   for(int r=0; r < p.nrow(); r++) {
+    if(use_norm[r]) {
+      out(r, _) = norm(nrm++, _);
+      for(int i=0; i < p.ncol(); i++) {
+        accept(r, i) = true;
+      }
+      continue;
+    }
+
     NumericVector kk = k(r, _);
     NumericVector nn = n(r, _);
     NumericVector mm = mean(r, _);
     for(int i = 0; i < p.ncol(); i++) {
       // safe not to replace out(r, i) because it's not used in this calculation
       double mmm = cond_mv_mean(out(r, _), mm, Q, i);
+
+      if(nn[i] == 0.0) {
+        out(r, i) = R::rnorm(mmm, 1/sqrt(Q(i, i)));
+        accept(r, i) = true;
+        continue;
+      }
       double ratio = one_m_ratio(out(r, i), proposal(r, i), kk[i], nn[i], mmm, Q(i, i));
       bool a = accept_reject(ratio);
       accept(r, i) = a;
@@ -90,10 +116,19 @@ NumericVector one_qt_proposal_ratio(double p, double k, double n, double mean, d
 
 
 // [[Rcpp::export]]
-NumericVector qt_binom(NumericVector p, NumericVector k, NumericVector n, NumericVector mean, NumericVector precision) {
+NumericVector qt_binom(NumericVector p, NumericVector k, NumericVector n, NumericVector mean, NumericVector precision,
+                       LogicalVector use_norm, NumericVector norm) {
   NumericVector out = clone(p);
   LogicalVector accept(p.size());
+
+  int nrm = 0;
   for(int i=0; i < p.size(); i++) {
+    if(use_norm[i]) {
+      out[i]= norm[nrm++];
+      accept[i] = true;
+      continue;
+    }
+
     NumericVector prop_ratio = one_qt_proposal_ratio(p[i], k[i], n[i], mean[i], precision[i]);
     bool a = accept_reject(prop_ratio[1]);
     accept[i] = a;
@@ -106,10 +141,21 @@ NumericVector qt_binom(NumericVector p, NumericVector k, NumericVector n, Numeri
 }
 
 // [[Rcpp::export]]
-NumericVector qt_binom_mv(NumericMatrix p, NumericMatrix k, NumericMatrix n, NumericMatrix mean, NumericMatrix Q) {
+NumericVector qt_binom_mv(NumericMatrix p, NumericMatrix k, NumericMatrix n, NumericMatrix mean, NumericMatrix Q,
+                          LogicalVector use_norm, NumericMatrix norm) {
   NumericMatrix out = clone(p);
   LogicalMatrix accept(p.nrow(), p.ncol());
+
+  int nrm = 0;
   for(int r=0; r < p.nrow(); r++) {
+    if(use_norm[r]) {
+      out(r, _) = norm(nrm++, _);
+      for(int i=0; i < p.ncol(); i++) {
+        accept(r, i) = true;
+      }
+      continue;
+    }
+
     NumericVector kk = k(r, _);
     NumericVector nn = n(r, _);
     NumericVector mm = mean(r, _);
@@ -117,6 +163,11 @@ NumericVector qt_binom_mv(NumericMatrix p, NumericMatrix k, NumericMatrix n, Num
       // safe not to replace out(r, i) because it's not used in this calculation
       double mmm = cond_mv_mean(out(r, _), mm, Q, i);
 
+      if(nn[i] == 0.0) {
+        out(r, i) = R::rnorm(mmm, 1/sqrt(Q(i, i)));
+        accept(r, i) = true;
+        continue;
+      }
       NumericVector prop_ratio = one_qt_proposal_ratio(out(r, i), kk[i], nn[i], mmm, Q(i, i));
       bool a = accept_reject(prop_ratio[1]);
       accept(r, i) = a;